Create LeaderboardTimeFilter
