+++ app/src/main/java/com/github/geohunt/app/ui/components/leaderboard/LeaderboardTitleBar.kt
@@ -1,21 +1,14 @@
 package com.github.geohunt.app.ui.components.leaderboard
 
-import androidx.compose.foundation.layout.*
-import androidx.compose.material.*
-import androidx.compose.material.icons.Icons
-import androidx.compose.material.icons.filled.ArrowDropDown
-import androidx.compose.runtime.*
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
 import androidx.compose.ui.unit.sp
 import com.github.geohunt.app.ui.theme.Lobster
-
-enum class LeaderboardTimeFilter(val text: String) {
-    ALL("All time"),
-    MONTHLY("Monthly"),
-    WEEKLY("Weekly")
-}
 
 /**
  * Draws the leaderboard title, as well as the time filter on the top-right
@@ -23,28 +16,6 @@ enum class LeaderboardTimeFilter(val text: String) {
  */
 @Composable
 fun LeaderboardTitleBar() {
-    // Whether the dropdown is expanded
-    var expanded by remember {
-        mutableStateOf(false)
-    }
-
-    // The currently chosen dropdown option
-    var filterOption by remember {
-        mutableStateOf(LeaderboardTimeFilter.ALL)
-    }
-
-    @Composable
-    fun DropdownOption(option: LeaderboardTimeFilter) {
-        DropdownMenuItem(
-            onClick = {
-                expanded = false
-                filterOption = option
-            }
-        ) {
-            Text(option.text)
-        }
-    }
-
     Row(
         Modifier.fillMaxWidth(),
         verticalAlignment = Alignment.CenterVertically
@@ -57,26 +28,6 @@ fun LeaderboardTitleBar() {
 
         Spacer(Modifier.weight(1.0f))
 
-        Box {
-            TextButton(
-                onClick = { expanded = !expanded }
-            ) {
-                Text(filterOption.text, color = Color.Black)
-                Icon(
-                    Icons.Default.ArrowDropDown,
-                    contentDescription = filterOption.text,
-                    tint = Color.Black
-                )
-            }
-
-            DropdownMenu(
-                expanded = expanded,
-                onDismissRequest = { expanded = false }
-            ) {
-                DropdownOption(LeaderboardTimeFilter.ALL)
-                DropdownOption(LeaderboardTimeFilter.MONTHLY)
-                DropdownOption(LeaderboardTimeFilter.WEEKLY)
-            }
-        }
+        LeaderboardTimeFilter()
     }
 }
